+++ agent/introspect/introspect_test.go
