+++ agent/introspect/package_test.go
