+++ cmd/jujud/main.go
@@ -27,13 +27,13 @@ import (
 
 	"github.com/juju/juju/agent/addons"
 	"github.com/juju/juju/agent/config"
+	"github.com/juju/juju/agent/introspect"
 	k8sexec "github.com/juju/juju/caas/kubernetes/provider/exec"
 	jujucmd "github.com/juju/juju/cmd"
 	agentcmd "github.com/juju/juju/cmd/jujud/agent"
 	"github.com/juju/juju/cmd/jujud/agent/agentconf"
 	"github.com/juju/juju/cmd/jujud/agent/caasoperator"
 	"github.com/juju/juju/cmd/jujud/dumplogs"
-	"github.com/juju/juju/cmd/jujud/introspect"
 	"github.com/juju/juju/cmd/jujud/run"
 	"github.com/juju/juju/core/arch"
 	"github.com/juju/juju/core/machinelock"
